requirements.txt
click>=8.0.0
python-dotenv>=0.19.0
openai>=1.0.0
anthropic>=0.7.0